Turn based combat system fixes, raycasting for instantiated real-time attacks
Added conditions for when to use enemy texture instead of ally texture.
Fixed some coding errors for the turn based system.
Modified dialogue code so it can be started at any point. Turn based combat pauses if a dialogue is active.
Added raycasting to get position for an instantiated attack that is separated from the player.