Interaction system complete(?), map loading update
The interaction system has been cleaned up and should be able to handle whatever it needs to, also added an interaction marker for clarity.
Added a loading screen which plays when loading maps, added an artificial loading timer to emulate what the experience would be like if the map was heftier to load.