Updates to both battle systems, update to dialogue system, added a function for spawning custom entities outside map hierearchy
Attacks with weapon an abilities now require an animation, added checks for local or instanced attacks as to allow both weapon attacks to spawn a separate entity (such as a ranged attack) and vice versa for abilities.
Turn based system works a little better, tested custom targetting and intertwining dialogue with battle states to good effect.
Also added a check for custom (boss) npcs on maps and allowed them to be spawned outside the map hierarchy.
Also fixed the loading screen not dissapearing and interfering with other ui elements.